refactor: extract product write operations into dedicated service
